Revert unnecessary cast

+++ tests/_files/source_with_multiline_constant_return.php
@@ -396,7 +396,7 @@ EOF;
     public function unaryMinusWithNotConstInTheMiddle(): float
     {
         return -
-        (int) (
+        (
             ''
             .
             phpversion()
